Add wallet unlocker component
Transactions made on the Ethereum network are signed using the wallet private key. For security purposes the private key is not stored in the local storage, and instead is temporarily stored in the memory.

Once the wallet unlocking dialog is shown and user enters the passphrase, we are attempting to unlock his v3 wallet (which can succed or not). In case of success we send a callback with unlocked wallet (the whole wallet object from ethereum-wallet).